Разложи весь код логики по классам
