+++ FileManager/FileManager/Program.cs
@@ -78,56 +78,56 @@ namespace FileManager
 
                 //    break;
                 //}
-                case 4:
-                    {
-                    Engine.DeleteCommandExecuter();
+                //case 4:
+                //    {
+                //    Engine.DeleteCommandExecuter();
 
-                    break;
-                }
-                case 5:
-                    {
-                    Engine.DeleteTreeCommandExecuter();
+                //    break;
+                //}
+                //case 5:
+                //    {
+                //    Engine.DeleteTreeCommandExecuter();
 
-                    break;
-                }
-                case 6:
-                    {
-                    Engine.MoveCommandExecuter(commandsStringArray[1], commandsStringArray[2]);
+                //    break;
+                //}
+                //case 6:
+                //    {
+                //    Engine.MoveCommandExecuter(commandsStringArray[1], commandsStringArray[2]);
 
-                    break;
-                }
-                case 7:
-                    {
-                    Engine.MakingDirectoryCommandExecuter();
+                //    break;
+                //}
+                //case 7:
+                //    {
+                //    Engine.MakingDirectoryCommandExecuter();
 
-                    break;
-                }
-                case 8:
-                    {
-                    Engine.RemoveDirectoryCommandExecuter(commandsStringArray[1], commandsStringArray[2]);
+                //    break;
+                //}
+                //case 8:
+                //    {
+                //    Engine.RemoveDirectoryCommandExecuter(commandsStringArray[1], commandsStringArray[2]);
 
-                    break;
-                }
-                case 9:
-                    {
-                    Engine.ShowAllSubdirectoriesAndFilesByPages();
+                //    break;
+                //}
+                //case 9:
+                //    {
+                //    Engine.ShowAllSubdirectoriesAndFilesByPages();
 
-                    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
+                //    PseudoConsoleUI.PrintDirectoryProrerties(Engine.drivesAndDirectories);
 
-                    break;
-                }
-                case 10:
-                    {
-                    PseudoConsoleUI.WriteHelp(Engine.HelpCommandExecuter());
+                //    break;
+                //}
+                //case 10:
+                //    {
+                //    PseudoConsoleUI.WriteHelp(Engine.HelpCommandExecuter());
 
-                    break;
-                }
-                case 11:
-                    {
-                    Engine.FilePropertiesCommandExecuter();
+                //    break;
+                //}
+                //case 11:
+                //    {
+                //    Engine.FilePropertiesCommandExecuter();
 
-                    break;
-                }
+                //    break;
+                //}
                 case 12:
                     {
                     Engine.ProcessCommandExecuter();
