+++ FileManager/FileManager/Temp/PseudoConsoleUI.cs
@@ -272,83 +272,88 @@ namespace FileManager
             Console.WriteLine();
         }
 
+        /// <summary>
+        /// Метод выводит на экран состояние текущей директории
+        /// </summary>
+        /// <param name="dirFiles">DrivesAndDirectories актуальное состояние программы</param>
+        public static void PrintDirectoryProrerties(DrivesAndDirectories dirFiles)
+        {
+            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+            Console.WriteLine();
+
+            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+            Console.ForegroundColor = ConsoleColor.DarkGray;
+
+            Console.Write($"Текущая директория: ");
+
+            Console.ForegroundColor = ConsoleColor.Magenta;
+
+            Console.Write($"{dirFiles.CuttentDirectory.FullName}");
+
+            Console.ForegroundColor = ConsoleColor.DarkGray;
+
+            Console.Write($" содержит: ");
+
+            Console.ForegroundColor = ConsoleColor.Magenta;
+
+            Console.Write($"{dirFiles.CuttentDirectory.GetFiles().Length}");
+
+            Console.ForegroundColor = ConsoleColor.DarkGray;
+
+            Console.Write($" файлов,");
+
+            Console.ForegroundColor = ConsoleColor.Magenta;
+
+            Console.WriteLine();
+
+            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+            Console.ForegroundColor = ConsoleColor.DarkGray;
+
+            Console.Write($"имеет атрибуты: ");
+
+            Console.ForegroundColor = ConsoleColor.Magenta;
+
+            Console.Write($"{dirFiles.CuttentDirectory.Attributes}");
+
+            Console.ForegroundColor = ConsoleColor.DarkGray;
+
+            Console.Write($", Время создания: ");
+
+            Console.ForegroundColor = ConsoleColor.Magenta;
+
+            Console.Write($"{dirFiles.CuttentDirectory.CreationTime:dd-MM-yyyy HH-mm-ss}");
+
+            Console.ForegroundColor = ConsoleColor.White;
+
+            Console.SetCursorPosition(0, Console.BufferHeight - 1);
+
+            Console.WriteLine();
+        }
+
+        #region For delete
+
         ///// <summary>
-        ///// Метод выводит на экран состояние текущей директории
+        ///// Метод выводит на экран Help - лист
         ///// </summary>
-        ///// <param name="dirFiles">DrivesAndDirectories актуальное состояние программы</param>
-        //public static void PrintDirectoryProrerties(DrivesAndDirectories dirFiles)
+        //public static void WriteHelp(List<string> list)
         //{
-        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
         //    Console.WriteLine();
 
-        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
+        //    foreach (string row in list)
+        //    {
+        //        Console.SetCursorPosition(0, Console.BufferHeight - 1);
 
-        //    Console.ForegroundColor = ConsoleColor.DarkGray;
-
-        //    Console.Write($"Текущая директория: ");
-
-        //    Console.ForegroundColor = ConsoleColor.Magenta;
-
-        //    Console.Write($"{dirFiles.CuttentDirectory.FullName}");
-
-        //    Console.ForegroundColor = ConsoleColor.DarkGray;
-
-        //    Console.Write($" содержит: ");
-
-        //    Console.ForegroundColor = ConsoleColor.Magenta;
-
-        //    Console.Write($"{dirFiles.CuttentDirectory.GetFiles().Length}");
-
-        //    Console.ForegroundColor = ConsoleColor.DarkGray;
-
-        //    Console.Write($" файлов,");
-
-        //    Console.ForegroundColor = ConsoleColor.Magenta;
-
+        //        Console.WriteLine(row);
+        //    }
         //    Console.WriteLine();
 
-        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-        //    Console.ForegroundColor = ConsoleColor.DarkGray;
-
-        //    Console.Write($"имеет атрибуты: ");
-
-        //    Console.ForegroundColor = ConsoleColor.Magenta;
-
-        //    Console.Write($"{dirFiles.CuttentDirectory.Attributes}");
-
-        //    Console.ForegroundColor = ConsoleColor.DarkGray;
-
-        //    Console.Write($", Время создания: ");
-
-        //    Console.ForegroundColor = ConsoleColor.Magenta;
-
-        //    Console.Write($"{dirFiles.CuttentDirectory.CreationTime:dd-MM-yyyy HH-mm-ss}");
-
-        //    Console.ForegroundColor = ConsoleColor.White;
-
-        //    Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-        //    Console.WriteLine();
         //}
 
-        /// <summary>
-        /// Метод выводит на экран Help - лист
-        /// </summary>
-        public static void WriteHelp(List<string> list)
-        {
-            Console.WriteLine();
+        #endregion
 
-            foreach (string row in list)
-            {
-                Console.SetCursorPosition(0, Console.BufferHeight - 1);
-
-                Console.WriteLine(row);
-            }
-            Console.WriteLine();
-
-        }
 
         /// <summary>
         /// Метод задает размеры окна и буфера консоли
